Remove flake8 exemption for wildcard imports
Wildcard imports undermine the capabilities of static analyzers
(namely `flake8`, but also pylint) and make it more difficult to
understand code.

Once a wildcard import is done, any apparently unbound name could be
taken from the wildcard import. It's not knowable without traversing
more modules.

They are *always* avoidable. In the extremely rare case that names
from a module or package must be rebound in another namespace, they
should all be imported explicitly and the `F401` flake8 rule (unused
imports) should be ignored. Given that these cases are often about
re-exporting names, we can even avoid this with the inclusion of an
`__all__` list or tuple.
In the case of `__init__.py` modules which exist solely for the
purpose of rebinding names as package-level members, we may prefer to
add a per-file-ignore for F401, or to simply always define an
`__all__` tuple. Either approach is satisfactory, and avoids
undermining `flake8`.

In several cases, the use of wildcard imports and allowance for F811
has masked redefinition of names with identical content --
presumably copy-paste mistakes in the past. These should not be
allowed.

When doing replacements within `funcx_endpoint`, I have manually
replicated the styles of `black`+`isort`, so that we will have a
marginally smaller diff when we run those tools on that package.