Adding a new kwarg FuncXClient(disable_requests_verify=Bool) to disable cert checking.
Some sites like quartz@LLNL appear to be doing some sort of SSL proxying that the
requests library is unable to verify. However, since curl is able to handle https connection,
this might just be a limitation of the `requests` library that. This option disables the
checking, but still raises a warning `InsecureRequestWarning: Unverified HTTPS request is being made to host 'api2.funcx.org'. Adding certificate verification is strongly advised`