Bound all funcx_sdk dependencies as appropriate
This ensures that no unused libraries are specified and that we
declare clear version requirements for those which are specified.

The only "questionable" bound is pinning an exact version of `dill`.
Because there is no changelog for dill and no publicly declared
versioning policy, pinning to an exact version seems like the best
choice.

Various `funcx_endpoint` requirements were declared in the SDK
requirements. The only one of these which is both really used AND
absent from the funcx_endpoint requirements data is `parsl`, so that
package is added there.

`packaging` is only used for its version parsing function. Although
`distutils.version` is deprecated, it is supported on all pythons
(including 3.10.0) and provides nearly-equivalent functionality for
version parsing. This allows us to remove a dependency.
`packaging` can be reintroduced at a later date if necessary.