Set safer endpoint dependency version bounds
Like the changes to `funcx-sdk`, this is meant to be a minimal touch,
removing unused things and setting strict bounds on the ones which
remain. Dependencies are *not* declared for funcx-sdk dependencies
which aren't also used directly in funcx-endpoint.

Several TODO items are listed in the requirements data. These are
mostly standard things, although there is a clear opinion expressed
here about `typer` not being worth the trouble for a `0.x` library.