Begin login config rewrite
Structure logging config more tightly, removing odd naming and highly
dynamic, imperative configuration of loggers.

The new structure is that there is a single method which can be used
to reconfigure logging on an as-needed basis (e.g. to turn off logging
to stdio).

Logging is configured at each entrypoint, and in one special case when
a daemon is being launched.

All loggers are initialized in module scope as

    log = logging.getLogger(__name__)

Any logger creation not matching this exact pattern is considered
incorrect and is fixed to match.