Remove internal RedisQueue object and dependents
RedisQueue is now provided by funcx-common if needed.

1. Remove `funcx_endpoint.queues`
2. Remove `funcx_endpoint/tests/integration/test_redis.py`
   (this just tests the removed subpackage)
3. Remove `funcx_endpoint.mock_broker`, which relies on RedisQueue

(3) is the most questionable of these three. However, `mock_broker`
does not appear to be used anywhere. Furthermore, even if a mocked
Forwarder object is needed for some reason, there doesn't appear to be
any reason that it should be provided by the `funcx-endpoint` package.