Update logging in Interchange
- Use setup_logging() to replace global logger object configured via
  set_file_logger()
- Cleanup CLI argument passing, as it relates to logging arguments
- Remove logging attributes from the Interchange object

This should, for the most part, not change the semantics of any code.
However, it is worth noting that it reduces the maxbytes on the
logfile from 256MB to 100MB. This could be adjusted back up by
allowing `setup_logging()` to take a parameter for maxbytes, but for
now this is not being done.