Remove logging helpers; fix "asyncio" logger
These were originally going to be left untouched, but the fact that
they were configuring logging for "asyncio" makes it hard to see a way
to salvage them.

The logger for "asyncio" _does not_ belong to funcx. We absolutely
should not be using that logger internally for funcx-related
activities.

Fix the references to the "asyncio" logger to use `__name__` as is
normal, and remove the imperative logging configuration helpers.