Use `FuncXClient.get_task` in get_result
`get_task` does not just get the task document and return it, so this
change was an unintentional behavioral change. Revert it.

Additionally, this commit includes the smoke test which flagged this
issue. The test submits a task which should return a pending status at
least once.

Right now, pending is encoded in the form of an exception.