Add support for 'reason' on FuncxAPIError
'reason' is a message field on the exception, meaning that when the
error is being parsed from JSON, the field is inspected and checked
for a string. If a string is found, it is set as `message`.

Notably, `message` is part of the default __str__ for a
GlobusAPIError, so this also puts `reason` into the string for an
exception.