Fix multiple loggers configured with same file
The Endpoint, Interchange, and Manager were all being launched with a
behavior that uses `<endpoint_dir>/endpoint.log` as the logfile.
Because logging is configured to use a RotatingFileHandler, the log
rotation in one process can "orphan" the loggers from the other two.
The result is lost log messages.

WatchedFileHandler would potentially be suitable in a future
implementation which merges these logfiles. However, it requires that
a distinct process handle rotation. This is therefore left open to
future refinement.

For an immediate fix, use distinct logfiles. This avoids any potential
for different processes clobbering one another's logs.