Remove use of ~/.funcx/endpoint.log
Instead of writing to this file, when logging setup is called without
a file, it sets up stderr stream logging. In order to make the output
more manageable in that case, put together a small custom formatter
which dispatches between formats depending on the log level of
records.

This way, any warnings or errors get lineno and other info, but we
don't "accidentally" use that in the output meant for end users.

Fix the way that logging config is mocked: instead of mocking over
the logfile being used just mock all logging setup.