Fix broken ScalingFailed exception construction
This tries to find the provider label inside self.config.provider,
which does not exist. In this interchange, the provider is
directly available as an attribute.

Tested by: modify my local kube provider to return None on
all submits, see that the issue #679 stack trace appears.
Make this change in this commit, and see that a ScalingFailed
correctly appears.

This addresses the first bullet point in issue #679.