Fix errors in docs
These are spotted if you fix things enough to run sphinx locally (as
seen in #691).

- title underline too short
- malformed code-block

In the case of the malformed code-block in the changelog, it also was
never subject to review. We should not be recommending that users
recreate complex logging utilities. Rather, show simple config and
point at the Python Logging HOWTO doc.