Ensure Interchange `setup_logging` runs post-fork
The `starter` function is used to launch the HTEX Interchange in a
`multiprocessing.Process` call. As such, it inherits the loggers from
the parent and was not fixing its logging.

In `starter`, call `setup_logging` to repoint logs. This does not
support passing through `debug`, but that's a limitation of the
interface today. It can be improved at a later date.