Add a tox config to test funcx-endpoint (#686)
Simplifies testing by codifying what is being done by CI in something
locally runnable.

Switch CI to using said config, and make sdk/endpoint jobs as uniform
as possible. In order for the safety check to work, put it in a
separate job that installs the funcx-endpoint directly.
Put the tox run of the endpoint into a matrix config which tests the
high and low python versions.