Enable mypy (passing on SDK, setup but failing on Endpoint) (#708)
* Setup mypy linting on sdk package

Minimal changes to make it pass. Annotate several un-annotated
variables. In the case of the WebSocketPollingTask, the internal
websockets client, `ws`, has been wrapped in a property to
guarantee that it is not-None when used.

* Setup runnable tox on endpoint (fails)

Configure the funcx_endpoint/tox.ini with a mypy build, but don't try
to run it in CI yet. This currently fails and can be fixed after the
renaming from the `rabbitmq` rearch branch.