Update release process: tox + tags (#694)
* Update release process: tox + tags

Add a `publish-release` tox env to both of the packages, so that
`tox -e publish-release` is sufficient to do a release in either one,
with no software dependencies for developers other than `tox`. This
also updates to using `build`, the pypa packaging tool for isolated
builds.

Replace `tag_and_release.sh` with a simpler `release.sh` script which
does the following:
- read the versions from both packages
- make sure the versions match
- make sure the version was added to the changelog
- tag with `git tag -s "$VERSION" -m "v$VERSION"`
- run `tox -e publish-release` in each package dir

The old script expected tox username. This is not correct -- e.g. your
credentials may be a pypirc with `__token__` as the username field for
modern pypi.org usage. Rely on twine to handle any pypi.org
credentials.

Most notably, this inverts the relationship between the release script
and version bumps. It is now an expectation that the version update is
handled (manually) prior to a release being started. In effect, this
means that someone could update both versions, run
`scriv --collect --edit` to update the changelog, commit the results
in a PR, and then open that PR prior to the release. That changeset
can then be reviewed *before* the release is done, which can then be
run either by the PR author or by the reviewer after the PR is merged
by running `./release.sh` on the main branch.

* Simplify release script further

Remove unnecessary args check and helper func.