Add _channel_closed condition to subscriber
On shutdown, TaskQueueSubscriber waits for the on_channel_closed
callback to fire, ensuring that channels are closed and no more
messages are being processed.
If the timeout (1 second, hardcoded for now) is reached, a warning is
logged. Add similar warning logging to the _cleanup_complete
event-wait.

Also, per feedback, remove `stop(block=False)` behavior, as it is only
useful in testing.