Cleanup TaskQueueSubscriber stop method (#712)
* Cleanup TaskQueueSubscriber stop method

- Instead of redefining `close()`, add a new method, `stop()`, which
  is responsible for terminating clenaly from the parent and invoking
  `close()` as one of the steps
- Use a ternary enum instead of `_closing = t/f` to describe a
  subscriber as "parent process", "running child" or "closing child".
  Call this the `status` and use it to track `closing`
- Define a `_shutdown()` method which handles clean teardown,
  including closing the connection to the MP queue and joining the
  background thread to ensure the queue has data flushed
- Update tests and fix one test

The enum could be done with an optional bool (true/false/null), but
then we'd need to explain that 'null' means 'parent process'. It's
clearer to use an enum.

Most of the test updates are just `close()` -> `stop()`. However, two
changes are interesting:

1. `stop(block=False); wait_until_stopped()` is used in one test to
   allow parallel shutdown of multiple subscriber objects following a
   typical "quit and join" pattern

2. test_subscriber_recovery was reusing the same disconnect event
   between multiple subscriber instances, but without clearing it
   inbetween

* Add _channel_closed condition to subscriber

On shutdown, TaskQueueSubscriber waits for the on_channel_closed
callback to fire, ensuring that channels are closed and no more
messages are being processed.
If the timeout (1 second, hardcoded for now) is reached, a warning is
logged. Add similar warning logging to the _cleanup_complete
event-wait.

Also, per feedback, remove `stop(block=False)` behavior, as it is only
useful in testing.