Improve test robustness for rabbitmq tests
Move all publisher and subscriber instantiation into dedicated fixture
functions. The fixture funcs put created objects into "teardown lists"
which have their relevant cleanup run at the end of each test in the
form of a finalizer function.

As a result, there is no danger of resource leaks around these tests
as every object is well accounted-for. The fixture funcs also fill in
numerous default values like connection parameters and the default
endpoint ID, so that tests are not required or expected to redundantly
define the same or similar information.

When publisher objects are created, the relevant queue_purge call is
made unless the helper is invoked with `queue_purge=False`, and
more generally the setup for these tests is consolidated into these
fixtures.

Tests are modified as necessary, and usually only as necessary. In a
few cases, there is relevant cleanup to the test logic included. In
particular, tests which use `FuncXSerializer` now use simple strings
instead. The serializer object is not under test in these cases, and
its use serves only to increase test runtime. Perf tests are mostly
unchanged, but now include asserts for the minimum acceptable
performance -- meaning that these are now tests which can fail,
whereas they were not proper tests beforehand.

The ResultQueueSubscriber object -- which exists primarily for testing
-- now has the full shutdown and wait behavior used in the
TaskQueueSubscriber. This ensures that shutdown can happen gracefully
during tests with no unexpected resource leaks.