Replace login as a built-in part of FuncXClient
Instead of login happening as part of client init (and only as part of
client init), login is now handled by an independent component called
the LoginManager. This encapsulates token storage and globus_sdk
client builders.

For the most part, the LoginManager replaces any notion that a
FuncXClient consumes authorizers for various internal clients
directly. Instead, it may be given an object (not necessarily a
LoginManager in the future) which is answerable for building usable
HTTP clients for various APIs which may be needed.