SC-11102 Implement task group rehydration.
By default, every instance of FuncXClient generates a fresh Task Group
ID.  In the happy-path case, the user processes their tasks and never
need know about it.  But in the case of a FuncXExecutor, access to this
allows a user to reattach to a previous session (for example, if the
researcher's script dies unexpectedly).  Reattaching looks like:

    saved_task_group_id = '<some_UUID_from_before>'
    fxc = FuncXClient(task_group_id=saved_task_group_id)
    fxe = FuncXExecutor(fxc)
    rehydrated_futures = fxe.reload_tasks()

At which point the futures are like before, and one can wait for their
results and exceptions:

    results, exceptions = [], []
    for fut in rehydrated_futures:
      try:
        results.append(fut.result(timeout=...))
    except Exception as exc:
        exceptions.append(exc)