Fix a race condition in interchange->manager task dispatch (#729)
The interchange dispatches tasks to managers based on how many tasks it thinks the manager needs.

This starts with the manager reporting how many task slots it has. However, the interchange may have already sent some tasks, so it separately tracks how many tasks have been sent, and reduces the number of tasks to be sent.

This prevents a task queue building up in a manager - such a task queue can lead to tasks building up in one manager, when another manager has free capacity.

However, at the other end of a task, when a task finishes, I have observed:

the manager now says it has capacity
the interchange knows there is still a task there, so does not allocate a task
the task result arrives
the manager sits idle until the next advertisement (which might be on the order of 30s sometimes)
This PR adds in a re-evaluation of the manager after a task result arrives, meaning that dispatch of a new task can happen right away.

In my kubernetes dev setup, this led to a run of 500 x 5s jobs on 6 manager-worker pods going from around 1000s wall time to 500s walltime.