Clarify what is tested in test_executor_shutdown
Based on the original PR for this work (#736), it appeared that this
test was checking the case in which the websocket connection exists.
In fact, because of which part of the executor is being mocked, it is
testing the no-connection case.

This can easily be checked by trying to access the connection (via its
property) and validating that a ValueError is raised.

If the code under test changes such that the mock no longer has the
desired effect, this will ensure it is flagged. Updated comments also
clarify exactly what is being tested.