Rename variables in FuncXExecutor for clarify
Previously tasks and task_id were used to refer to a "thing that will
probably become a funcx task later, but for now is just a Future".

The id is renamed future_id.

The _tasks dictionary is renamed to be consistent with _function_future_map,
to _counter_future_map.

The type of _counter_future_map is fixed.