Make TaskSubmissionInfo a class for better types
Rather than a namedtuple (which works fine), declare this as an
absolutely minimal class with a small __repr__ , which we'll go ahead
and test too.

The reason is that this improves the behavior of the type annotations
for that object. Its attribute types are known, and the parameters to
its constructor are marked as required.