Convert executor internal message objects from dict to dedicated object (#745)
* Convert the task submission dict to namedtuple

This allows type annotations to show the type of this object.
It simplifies packing/unpacking and lets those operations type-check.

This also handily hints that the data in question is purely used on
the local side, in-memory, and isn't about to be sent to the service
as-is.

* Cleanup conditions in executor batch submission

Minor cleanup:
- replace 'while True: if X: break' with 'while X'
- move conditional check outside of `_submit_tasks`

* Make TaskSubmissionInfo a class for better types

Rather than a namedtuple (which works fine), declare this as an
absolutely minimal class with a small __repr__ , which we'll go ahead
and test too.

The reason is that this improves the behavior of the type annotations
for that object. Its attribute types are known, and the parameters to
its constructor are marked as required.