Cleanup for faster client init (#753)
The basic goal here is (much) faster test runs. To achieve this, three
changes are made to FuncXClient:

1. Add a new argument, `do_version_check=False` to skip the version
   check callout

2. Avoid doing a userinfo callout to Globus Auth by removing the need
   for the user's ID from the SearchHelper object

3. Lazily initialize the SearchHelper object only when accessed as
   part of a path towards potential deprecation and removal

The tests should have mocked this functionality sufficiently for
things to run very quickly, but for some reason they were slow.
Perhaps the activities of the SearchHelper object were to blame, or
perhaps something was amiss with the mocks. Rather than worry overmuch
about exactly how the time was being spent, this "cuts through the
noise" by trying to make client initialization faster in general, with
tests only being one of the places which benefit.

Additionally, tests are now initializing FuncXClient objects with
`do_version_check=False, use_offprocess_checker=False` to attempt to
reduce external interactions from this code to nearly-zero.

`openid_authorizer` as an argument is no longer used. Rather than
doing nothing, this takes the extra step of marking it as deprecated
so that it can be removed in the following release.