Remove distutils and reorganize version check (#754)
* Remove distutils and reorganize version check

- Remove distutils (deprecated in py3.10). Write a small version
  parsing helper instead (can use `packaging` in the future)
- set do_version_check=False in the endpoint code
- remove check_endpoint_version as a client init argument
- version_check can now accept the endpoint version as an input
  (instead of that being automatic from local package versioning
  information), removing the funcx->funcx-endpoint import path

* Add tests for custom version parsing

* Make version parsing handle '-dev' as '.dev0'

The way that 'packaging' handles a '-dev' suffix is to mark that the
same as '.dev0', so we will do the same. It makes no semantic
difference since we're not trying to support all version forms.

* Add explicit version check to endpoint configure

This may not be necessary, but use the client to check versions during
setup. Also use 'use_offprocess_checker=False' to get slightly quicker
starts.

* Fix the type annotation for version 'pre'

After the change to `('dev', 0)`, the type here needed adjustment.

* Revert changes incompatible with last release

`funcx-endpoint` can't change such that it is not compatible with the
last release of `funcx` until a release is done.

Because `FuncXClient` accepts arbitrary `kwargs`, it's okay to pass
`do_version_check` and `check_endpoint_version`. Each of these will be
ignored on a relevant version of `funcx`.

* Remove check_endpoint_version=True usage

`funcx-endpoint start` was passing `check_endpoint_version=True` which
is presently being deprecated and removed.