Refactor _update_task_table.
The major update here is mostly for robustness.  In the context of
reloading tasks, it's possible to request a task from upstream that
doesn't exist -- and thus returned from the API as 'Failed' but without
an exception.  Catch that case by refactoring logic to better understand
"pending" and to always create the exception.