Update smoke tests to handle login manager
In the event that the smoke tests are running against old code, they
need to pass authorizers. On the new version, they will pass a login
manager object to simulate a logged-in state.

To handle this, try to import the LoginManagerProtocol and use it for
a runtime check to ensure we've implemented everything necessary. If
the import fails, failover to the current (old) behavior.