Replace login as a built-in part of FuncXClient (#724)
* Replace login as a built-in part of FuncXClient

Instead of login happening as part of client init (and only as part of
client init), login is now handled by an independent component called
the LoginManager. This encapsulates token storage and globus_sdk
client builders.

For the most part, the LoginManager replaces any notion that a
FuncXClient consumes authorizers for various internal clients
directly. Instead, it may be given an object (not necessarily a
LoginManager in the future) which is answerable for building usable
HTTP clients for various APIs which may be needed.

* Add LoginManagerProtocol and update changelog

The LoginManagerProtocol defines what the requirements are for an
alternative login manager implementation to be passed to a
FuncXClient.

Changelog updates make the description of the login manager change
more accurate after a rebase.

* Update smoke tests to handle login manager

In the event that the smoke tests are running against old code, they
need to pass authorizers. On the new version, they will pass a login
manager object to simulate a logged-in state.

To handle this, try to import the LoginManagerProtocol and use it for
a runtime check to ensure we've implemented everything necessary. If
the import fails, failover to the current (old) behavior.

* Use typing_extensions on py3.7 for Protocol

* Pass env from client to default LoginManager

`environment:str|None` is added in several components and passed through
so that `environment="dev"` will use the appropriate tokens.

* Fix mocking of login in tests