Don't redundantly do the work.
`pending_futures` only "gets smaller" and `unknown_futures` only "gets
bigger" -- in relation to `pending_futures` -- for the duration of the
method execution.  Thus, it is only beneficial to check if unknown
task_ids are now known when we first get new pending futures to handle.

While here correct a bug in order to make testing easier -- the case of
"something going wrong" with parsing of the datagram.  The general
exception was logged and then ignored.  Fix that by setting the future
exception.