Cleanup config for running endpoint tests
1. Move all dead tests to an "archived" directory for later perusal or
   deletion (as the team deems fit). Mark `tests/archived` in pytest
   `norecursedirs`
2. Make tox run the testsuite with a simple pytest invocation
3. For clean coverage reporting, make tests which generate a python
   "config.py" module do so in a tmp dir

Regarding (3), this was revealed by the other changes. The generated
module does not play nicely with `coverage` and this was somehow not
being flagged before.
I would like to later revisit the use of a python module for config at
all, but this commit is limited to only the necessary changes.