Remove inappropriate use of 'utils' from funcx_sdk
Many projects find need of a utility module or subpackage, serving
as a place for various simple helpers or wrappers over the stdlib to
provide nicer APIs. However, funcx_sdk is using it in three places
where it is not appropriate or accurate to use the name 'utils':

1. In unit tests, a module named utils for shared imports

pytest fixtures provide smoother code sharing without requiring
that the test modules be in an importable package.

2. `funcx.sdk.utils`, a subpackage which only contains the Batch
   object

If this object is going to be used as a helper for building requests,
it is not an internal utility, but a part of the API surface for the
FuncXClient. Elevate it from `funcx.sdk.utils.batch` to
`funcx.sdk.batch`.

3. `funcx.utils` only contains error classes and one error handling
   helper

This is clearly not "utils" at all. The name of this subpackage is
more properly `funcx.errors`, `funcx.exceptions`, or similar. Rename
to `funcx.errors` and provide an `__init__.py` which defines all of
the names which are importable from `funcx.errors` for good
encapsulation.

The "utils" name has knock-on effects downstream. For example,
there may be client code which uses

    from funcx.utils.errors import ...

which now would be fixed to

    from funcx.errors import ...

As such, this deserves a changelog entry, included here.

All usage sites in this repo -- funcx-endpoint and smoke_tests -- are
updated to try-except(ImportError) to handle both potential paths for
error class imports.