Convert from 'typer' to 'click' (#787)
* Convert from 'typer' to 'click'

All use of typer is replaced with click.

Simple CLI application configuration, support for `--debug` on any
command and various other goodies included. All significant features
are maintained except for the autocompletion of endpoint names. This
could be reintroduced for some commands by means of `click.Choice` as
an option type if desired. For now this is omitted as a future
feature.

The following CLI interface changes were made:
- The optional "name" argument to various commands is replaced with
  `-n, --name` as an option
- `--debug` is supported either on the root command or on the
  subcommands
- `-h` and `--help` are supported for the help opt
- `funcx-endpoint --version` is replaced with `funcx-endpoint version`

* helm boot script for endpoint uses `--endpoint-id`

This was changed away from using `--endpoint-id`, but the rewritten
CLI parsing makes that option required. Change it back.

* Make `--name` into an optional positional arg

The change to `--name` is reverted here. Instead of an option, it's an
optional positional.