Refactor result sending to use messagepack
Using the 'messages_compat' module, convert result dicts (errors and
successful results) to Result message objects. The entire internal
pipeline for results in the endpoint relies on them being pickled
dicts, so any manipulations to the data being sent are done first on
the contents of that dict, and then at "translation time" at the end.
Most notably, the exception object is now passed back in the form of a
string combined with the error_details fields as a tuple.

When sending to the result-processor, these data can be composed into
a messagepack Result message object, serialized, and sent.

On the receiving end in FuncXClient, some adjustment is needed as
well. The arriving message does not contained a serialized exception
on errors, so the error string (if present) is unpacked into a raw
exception object which is raised. The error __str__ definition is an
indented string, so when shown in a traceback it appears in a coherent
and identifiable way as "part of the trace".

In the case of non-exception data, the data is serialized as it was
before.