SC-14129 Do not store connection url.
This is nominally sensitive data -- if ephemeral -- and isn't required to be
written at all.  Everytime an endpoint starts, it registers with the webserver.
At that time, the webserver will provide *new* RabbitMQ credentials.
Consequently, there is only a need to keep them in memory long enough to open
the connection.