Fixup mypy (#805)
* Update mypy deps with typeshed

Add types-retry and types-requests to get stubs from typeshed. Also,
pin mypy version.

* Fix mypy failures on "messages"

* Add annotations to Kubernetes provider

* Fix type annotations in logging_config

The trace level addition hits a bunch of attr definition errors.
Ignore all of these, as mypy will not be able to understand the
addition of levels to `logging`.

* Fix annotations of multiprocessing.Event

This is actually a method, not a class. The fix is to import the
underlying class under a different name and use it for annotations.

* Update try-except import of MaxResultSizeExceeded

This causes a mypy error and is marked as no longer needed.
For a short while, we will need to try both names of this error
(because it was renamed), but there is no need to define the class
inline.

* Fix annotations in worker_map

* Fix type annotations in ResultQueuePublisher

* Fix annotations in TaskQueuePublisher

* Fix annotations for htex manager

* Fix annotations in htex interchange

Except for those which require semantic changes.

* Fix type of htex interchange pending_task_queue

And remove a method which is not compatible with the observed type of
the object. Also remove the `tasks` set, which is never used and
therefore can't be accurately typed.

* Update github actions and `make lint` to run mypy