SC-14129 Use webserver provided queue information
The endpoint's interaction with the RMQ server is governed by the webserver's
setup (e.g., RMQ credentials, which queues to read/write, queue policies), so
do not hard code exchange and queue information -- the webserver provides this
information in the endpoint registration response.

Accordingly, address the respective integration tests.  While in the tests,
address a bug whereby stale messages are left in the results queue after a test
run, set all of the test-created queues to auto expire, and use test-specific
queue names (to aid in debugging; e.g. reading RMQ logs)