SC-14129 Address test-introduced race-condition.
I was unable to recreate the issue locally, but GitHub's CI infrastructure hit
it almost every time: the for-testing-purposes-only ResultQueueSubscriber needs
to be completely initialized before the test is valid and may continue.
Accordingly, introduce a barrier specifically for validity.  (And add newline
to log for easier human eyeball parsing.)