Refine and test FuncXWorker.execute_task
The unpacking of the user function now supports a
messagepack.message_types.Task object in addition to the "legacy"
internal message format. This accommodates testing better, as it
provides a newer and clearer interface for constructing tasks to send
directly to the worker.

Test execute_task on three basic cases:
- hello_world (success)
- KeyError (failure)
- MaxResultSizeExceeded (failure)

In each case, the result should be constructed and returned as a dict
for now. In the future it will likely be refined into a `Result`
message object.

The testing ensures that the traceback munging done with
`tb.tb_next.tb_next` behaves appropriately and removes the funcx
methods from the stackframes included in the error data. This ensures
that users see tracebacks starting with their own code, not with the
funcx wrapping.