Convert outgoing results to messagepack (#814)
* Refactor result sending to use messagepack

Using the 'messages_compat' module, convert result dicts (errors and
successful results) to Result message objects. The entire internal
pipeline for results in the endpoint relies on them being pickled
dicts, so any manipulations to the data being sent are done first on
the contents of that dict, and then at "translation time" at the end.
Most notably, the exception object is now passed back in the form of a
string combined with the error_details fields as a tuple.

When sending to the result-processor, these data can be composed into
a messagepack Result message object, serialized, and sent.

On the receiving end in FuncXClient, some adjustment is needed as
well. The arriving message does not contained a serialized exception
on errors, so the error string (if present) is unpacked into a raw
exception object which is raised. The error __str__ definition is an
indented string, so when shown in a traceback it appears in a coherent
and identifiable way as "part of the trace".

In the case of non-exception data, the data is serialized as it was
before.

* Fix client unit tests to work after result changes

The result message format and contents have changed. As such, the
behavior has also changed a bit. Some test adjustments were needed. In
particular, a parametrized test encoding multiple behaviors has been
broken up into several tests to make the test code simpler, and the
error checking tests include coverage for a set of ValueErrors which
can be triggered on bad inputs to the function under test.

To handle these changes, repair and improve existing test of FuncXWorker.
Fix registration error inspection, don't mock sys.exit, don't do weird
stuff in the worker class itself to accommodate tests (return after
sys.exit), convert the test class to a single test function.

* Refine and test FuncXWorker.execute_task

The unpacking of the user function now supports a
messagepack.message_types.Task object in addition to the "legacy"
internal message format. This accommodates testing better, as it
provides a newer and clearer interface for constructing tasks to send
directly to the worker.

Test execute_task on three basic cases:
- hello_world (success)
- KeyError (failure)
- MaxResultSizeExceeded (failure)

In each case, the result should be constructed and returned as a dict
for now. In the future it will likely be refined into a `Result`
message object.

The testing ensures that the traceback munging done with
`tb.tb_next.tb_next` behaves appropriately and removes the funcx
methods from the stackframes included in the error data. This ensures
that users see tracebacks starting with their own code, not with the
funcx wrapping.

* Fix mypy warnings about FuncXWorker

The traceback traversal is flagged because `tb_next` could be `None`.
Rectify this with a safer helper.

Cast `sys.exc_info`'s return type to non-optional since it's executing
inside a handler.

Ignore attr access on an object whose type is not knowable to mypy.

* Fix how internal errors render in Result messages

The exception detail for a MaxResultSizeExceeded or
CouldNotExecuteUserTaskError should *not* contain a full trace because
 - the trace would not be particularly useful
 - the trace context will show funcx internals, not user code

This was causing a test to fail after traceback traversal got fixed.